+++ src/store/modules/reducers.js
@@ -1,0 +1,9 @@
+import { combineReducers } from 'redux';
+
+import user from './user/reducers';
+
+const reducers = combineReducers({
+  user,
+});
+
+export default reducers;